% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection.R
\name{bw_select_cv_univariate}
\alias{bw_select_cv_univariate}
\title{Cross-validation for univariate distributions}
\usage{
bw_select_cv_univariate(x, tol = 10^(-3))
}
\arguments{
\item{x}{The vector of data points.}

\item{tol}{The absolute tolerance in the optimization}
}
\description{
Uses cross-valdiation to find the optimal bandwidth for a univariate locally
Gaussian fit
}
\details{
This function looks at a vector of observations and estimates the bandwidth that
minimizes the Kullback-Leibler distance between the true and the estimated densities.
}
