% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_bw_method}
\alias{check_bw_method}
\title{Check bw method}
\usage{
check_bw_method(bw_method)
}
\arguments{
\item{bw_method}{Check if equal to "cv" or "plugin"}
}
\description{
Checks that the bandwidth method is one of the allowed values, currently "cv"
or "plugin".
}
