% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_bw_trivariate}
\alias{check_bw_trivariate}
\title{Check bandwidth vector}
\usage{
check_bw_trivariate(bw)
}
\arguments{
\item{bw}{The bandwidth vector to be checked}
}
\description{
Checks that the bandwidth vector supplied to the bivariate density function
is a numeric vector of length 3.
}
