% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_est_method}
\alias{check_est_method}
\title{Check estimation method}
\usage{
check_est_method(est_method)
}
\arguments{
\item{est_method}{Check if equal to a valid value}
}
\description{
Checks that the estimation method is one of the allowed values, currently
"1par", "5par" and "5par_marginals_fixed".
}
