% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_lg}
\alias{check_lg}
\title{Check that an object has class "lg"}
\usage{
check_lg(check_object)
}
\arguments{
\item{check_object}{The object to be checked}
}
\description{
Check that an object has class "lg"
}
