% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_conditional_independence.R
\name{ci_test}
\alias{ci_test}
\title{Test for conditional independence}
\usage{
ci_test(lg_object, n_replicates = 1000, h = function(x) x^2,
  return_replicated_test_statistics = TRUE, return_replicated_data = FALSE,
  return_auxiliary_info = FALSE)
}
\description{
Test for conditional independence using local Gaussian correlations.
}
\details{
This is the main function for performing a test for conditional independence
between two variables given a (set of) variable(s). The function takes in an
lg-object as produced by the main \code{lg}-function,
}
