% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{clg}
\alias{clg}
\title{The locally Gaussian conditional density estimator}
\usage{
clg(lg_object, grid = NULL, condition = NULL)
}
\arguments{
\item{lg_object}{An object of type \code{lg}, as produced by the \code{lg}-function}

\item{grid}{A matrix of grid points, where we want to evaluate the density estimate}

\item{condition}{A vector with conditions for the variables that we condition upon}
}
\description{
Estimate a conditional density function using locally Gaussian approximations.
}
\details{
This function is the coditional version of the locally Gaussian density estimator
(LGDE), described in Otneim & Tjøstheim (2017) [Proper citation needed!]. The
function takes as arguments an \code{lg}-object as produced by the main \code{lg}-
function, a grid of points where the density estimate should be estimated, and a
set of conditions.

The variables must be sorted befor they are supplied to this function. It will always
assume that the free variables come before the conditioning variables.
}
