% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{corplot}
\alias{corplot}
\title{Plot unconditional local correlations}
\usage{
corplot(dlg_object, pair, plot_map = TRUE, gaussian_scale = FALSE,
  plot_colormap = TRUE, plot_obs = FALSE, plot_labels = TRUE,
  plot_legend = FALSE, alpha_tile = 0.5, alpha_point = 0.8,
  low_color = "blue", high_color = "red", break_int = 0.2,
  label_size = 3, font_family = "sans", point_size = NULL, xlab = NULL,
  ylab = NULL, rholab = NULL, main = NULL, subtitle = NULL)
}
\description{
blabla
}
\details{
blablajkaklajkl
}
