% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{dlg}
\alias{dlg}
\title{The locally Gaussian density estimator (LGDE)}
\usage{
dlg(lg_object, grid = NULL)
}
\arguments{
\item{lg_object}{An object of type \code{lg}, as produced by the \code{lg}-function}

\item{grid}{A matrix of grid points, where we want to evaluate the density estimate}
}
\description{
Estimate a mulivariate density function using locally Gaussian approximations
}
\details{
This function does multivariate density estimation using the locally Gaussian
density estimator (LGDE), that was introduced by Otneim & Tjøstheim (2016)
[Proper citation needed!]. The function takes as arguments an \code{lg}-object
as produced by the main \code{lg}-function, and a grid of points where the density
estimate should be estimated.
}
