% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{dlg_bivariate}
\alias{dlg_bivariate}
\title{Bivariate density estimation}
\usage{
dlg_bivariate(x, eval_points, bw, est_method,
  tol = .Machine$double.eps^0.25/10^4, run_checks = TRUE)
}
\description{
\code{dlg_bivariate} returns the locally Gaussian density estimate on a given grid.
}
\details{
This function serves as the backbone in the body of methods concerning local
Gaussian correlation. It takes a bivariate data set, \code{x}, and a bivariate
set of grid points \code{eval_points}, and returns the bivariate, localy Gaussian
density estimate in these points. We also need a vector of bandwidths, \code{bw},
with two elements, and an estimation method \code{est_method}

TODO: Add proper default values when they become available. Document further.
}
