% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{dlg_bivariate_loop_helpfunc}
\alias{dlg_bivariate_loop_helpfunc}
\title{Help function for concise looping in sequential bivariate density estimation}
\usage{
dlg_bivariate_loop_helpfunc(i, x, x0, lg_object, marginal_estimates, pairs)
}
\arguments{
\item{i}{Indicate pair number}

\item{x}{The bivariate data set}

\item{x0}{Evaluation points}

\item{lg_object}{The \code{lg}-object calculated for this data set}

\item{marginal_estimates}{The marginal density estimates}

\item{pairs}{Matrix indicating the pairs of variables.}
}
\description{
Function that is used in accomodating paralell computing of locally Gaussian
density estimation. This function is only used internally.
}
