% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{dlg_marginal}
\alias{dlg_marginal}
\title{Marginal density estimation}
\usage{
dlg_marginal(x, bw = 1, eval_points = seq(quantile(x, 0.01), quantile(x,
  0.99), length.out = grid_size), grid_size = 15)
}
\arguments{
\item{x}{The data vector.}

\item{bw}{The bandwidth (a single number).}

\item{eval_points}{The grid where we want to evaluate the density. Chosen
suitably if not provided, with length equal to grid_size.}

\item{grid_size}{Number of grid points if grid is not provided.}
}
\description{
Function that estimates a univariate density estimation by local Gaussian
approximations
}
\details{
This functio is mainly mean to be used as a tool in multivariate analysis
as away to obtain the estimate of a univariate (marginal) density function,
but it can of course be used in general to estimate univariate densities.
}
