% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimation.R
\name{dlg_marginal_wrapper}
\alias{dlg_marginal_wrapper}
\title{Marginal estimates for multivariate data}
\usage{
dlg_marginal_wrapper(data_matrix, eval_matrix, bw_vector)
}
\arguments{
\item{data_matrix}{The matrix of data points. One column constitutes an
observation vector.}

\item{eval_matrix}{The matrix of evaluation points. One column constitutes
a vector of grid points.}

\item{bw_vector}{The vector of bandwidths, one element per component.}
}
\description{
Estimates the marginal locally Gaussian parameter estimates for a multivariate
data set
}
\details{
This function takes in a matrix of observations, a matrix of evaluation points
and a vector of bandwidths, and does the marginal estimation pairwise. This
function assumes that the data and evaluation points are organized column-wise
in matrices, and that the bandwidth is found in the corresponding element in
the bandwidth matrix.
}
