% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{dmvnorm_wrapper_single}
\alias{dmvnorm_wrapper_single}
\title{Wrapper for \code{dmvnorm} - single point}
\usage{
dmvnorm_wrapper_single(x1, x2, mu_1, mu_2, sig_1, sig_2, rho)
}
\arguments{
\item{x1}{The first component of the evaluation point}

\item{x2}{The second component of the evaluation point}

\item{mu_1}{The first expectation}

\item{mu_2}{The second expectation}

\item{sig_1}{The first standard deviation}

\item{sig_2}{The second standard deviation}

\item{rho}{The correlation}
}
\description{
Function that evaluates the bivariate normal in a single point
}
