% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_conditional_independence.R
\name{generate_replicates}
\alias{generate_replicates}
\title{Generate replicates under the null hypothesis for conditional independence test}
\usage{
generate_replicates(x, n_replicates, h)
}
\description{
Generate replicates under the null hypothesis for conditional independence test
}
