% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{gradient}
\alias{gradient}
\title{Auxiliary function for calculating the asymptotic standard deviations for the
local Gaussian correlations}
\usage{
gradient(sigma, sigma_k)
}
\arguments{
\item{sigma}{sigma}

\item{sigma_k}{sigma_k}
}
\description{
Auxiliary function for calculating the asymptotic standard deviations for the
local Gaussian correlations
}
