% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_tests.R
\name{ind_teststat}
\alias{ind_teststat}
\title{Function that calculates the test statistic in the independence tests.}
\usage{
ind_teststat(x_replicated, lg_object, S, h)
}
\arguments{
\item{x_replicated}{A sample.}

\item{lg_object}{An lg-object.}

\item{S}{Integration area, see \code{?ind_test}.}

\item{h}{h-function for test statistic, see \code{?ind_test}.}
}
\description{
This is an auxiliary function used by the independence tests.
}
