% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{lik_1par}
\alias{lik_1par}
\title{1-parameter likelihood function for bivariate data}
\usage{
lik_1par(rho, m1, m2, m3, m4, x1_0, x2_0, h1, h2)
}
\arguments{
\item{rho}{rho}

\item{m1}{m1}

\item{m2}{m2}

\item{m3}{m3}

\item{m4}{m4}

\item{x1_0}{x1_0}

\item{x2_0}{x2_0}

\item{h1}{h1}

\item{h2}{h2}
}
\value{
The likelihood for the one-parameter case.
}
\description{
\code{lik_1par} returns the likelihood function
}
