% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{make_C}
\alias{make_C}
\title{Auxiliary function for calculating the asymptotic standard deviations for the
local Gaussian correlations}
\usage{
make_C(r, pairs, p)
}
\arguments{
\item{r}{r}

\item{pairs}{pairs}

\item{p}{p}
}
\description{
Auxiliary function for calculating the asymptotic standard deviations for the
local Gaussian correlations
}
