% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{maximize_likelihood_1par}
\alias{maximize_likelihood_1par}
\title{1-parameter likelihood maximization function for bivariate data}
\usage{
maximize_likelihood_1par(grid_point, x1, x2, h1, h2, tol)
}
\arguments{
\item{grid_point}{Grid point}

\item{x1}{The first data vector}

\item{x2}{The second data vector}

\item{h1}{The first bandwidth}

\item{h2}{The second bandwidth}

\item{tol}{The tolerance used by the \code{optim}-function}
}
\value{
The maximized likelihood for the one-parameter case.
}
\description{
Optimization of the local likelihood function.
}
