% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{mvnorm_eval}
\alias{mvnorm_eval}
\title{Evaluate the multivariate normal}
\usage{
mvnorm_eval(eval_points, loc_mean, loc_sd, loc_cor, pairs)
}
\arguments{
\item{eval_points}{A matrix of grid points}

\item{loc_mean}{A matrix of local means, one row per grid point, one column
per component}

\item{loc_sd}{A  matrix of local standard deviations, one row per grid point,
one column per component}

\item{loc_cor}{A matrix of local correlations, one row per grid point, on
column per pair of variables}

\item{pairs}{A data frame specifying the components that make up each pair,}
}
\description{
Function that evaluates the multivariate normal distribution with local
parameters
}
\details{
Takes in a grid, where we want to evaluate the multivariate normal, and in
each grid point we have a new set of parameters.
}
