% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_correlation.R
\name{partial_cor}
\alias{partial_cor}
\title{Calculate the local Gaussian partial correlation}
\usage{
partial_cor(lg_object, grid = NULL, condition = NULL)
}
\description{
A function that calculates the local Gaussian partial correlation for a pair
of variables, given the values of some conditionaing variables.
}
\details{
Explanation....
}
