% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{trans_normal}
\alias{trans_normal}
\title{Transform the marginals of a multivariate data set to standard normality based
on the logspline density estimator.}
\usage{
trans_normal(x)
}
\arguments{
\item{x}{The data matrix, one row per observation.}
}
\value{
A list containing the transformed data ($transformed_data), and a function
($trans_new)  that can be used to transform grid points and obtain normalizing
constants for use in density estimation functions
}
\description{
Transform the marginals of a multivariate data set to standard normality based
on the logspline density estimator.
}
