% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_data}
\alias{check_data}
\title{Check the data and grid}
\usage{
check_data(x, dim_check = NA, type)
}
\arguments{
\item{x}{Data or grid}

\item{dim_check}{How many columns do we expect?}

\item{type}{Is it the "grid" or "data" for use in error messages.}
}
\description{
Checks that the data or grid provided is of the correct form. This function
is an auxiliary function that can quickly check that a supplied data set or
grid is a matrix or a data frame, and that it has the correct dimension, as
defined by the \code{dim_check} parameter. The \code{type} argument is simply
a character vector "data" or "grid" that is used for printing error messages.
}
