% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{lik_1par}
\alias{lik_1par}
\title{1-parameter likelihood function for bivariate data}
\usage{
lik_1par(rho, m1, m2, m3, m4, x1_0, x2_0, h1, h2)
}
\arguments{
\item{rho}{The local Gaussian correlation, with respect to which we optimize}

\item{m1}{Empirical mean 1, used in the optimization}

\item{m2}{Empirical mean 2, used in the optimization}

\item{m3}{Empirical mean 3, used in the optimization}

\item{m4}{Empirical mean 4, used in the optimization}

\item{x1_0}{Grid point 1}

\item{x2_0}{Grid point 2}

\item{h1}{Bandwidth 1}

\item{h2}{Bandwidth 2}
}
\value{
The likelihood for the one-parameter case.
}
\description{
Used for optimization, only used internally.
}
