% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{maximize_likelihood_1par}
\alias{maximize_likelihood_1par}
\title{1-parameter likelihood maximization function for bivariate data}
\usage{
maximize_likelihood_1par(grid_point, x1, x2, h1, h2, tol)
}
\arguments{
\item{x1}{x1}

\item{x2}{x2}

\item{h1}{h1}

\item{h2}{h2}

\item{tol}{tol}

\item{rho}{grid_point}
}
\value{
The maximized likelihood for the one-parameter case.
}
\description{
1-parameter likelihood maximization function for bivariate data
}
