% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{u}
\alias{u}
\title{Auxiliary function for calculating the local score function u}
\usage{
u(z1, z2, rho)
}
\arguments{
\item{z1}{z1}

\item{z2}{z2}

\item{rho}{rho}
}
\description{
Auxiliary function for calculating the local score function u
}
\details{
This function is used to estimate the asymptotic variance of the estimates.
}
